+++ scripts/solve.mjs
@@ -1,41 +1,3 @@
-import { readFile, writeFile } from "fs/promises";
-import { basename } from "path";
-import { patch } from "./utils/patch.mjs";
-import { fileURLToPath } from "node:url";
-import { findFile } from "./utils/findFile.mjs";
+import { patchFile } from "./utils/patchFile.mjs";
 
-const sourceFileNameWithExt = await findFile(process.argv[2], "patch");
-if (!sourceFileNameWithExt) {
-  console.log(
-    `No file matching ${process.argv[2]} found in the exercises folder.`
-  );
-  process.exit(1);
-}
-
-const nameBase = basename(sourceFileNameWithExt, ".patch");
-
-const patchFilePath = fileURLToPath(
-  new URL(`../patch/${nameBase}.patch`, import.meta.url)
-);
-const patchFile = await readFile(patchFilePath, "utf-8").catch(() => {
-  console.error(`No patch file found under patch/${nameBase}.patch`);
-  process.exit(1);
-});
-
-const sourceFilePath = fileURLToPath(
-  new URL(`../exercises/${nameBase}.wat`, import.meta.url)
-);
-const sourceFile = await readFile(sourceFilePath, "utf-8");
-
-try {
-  const patchedContent = patch(patchFile, sourceFile);
-
-  await writeFile(sourceFilePath, patchedContent);
-
-  console.log(`Patch applied successfully to: ${nameBase}`);
-} catch (e) {
-  console.error(`Error applying patch: ${e.message}`);
-  console.info(
-    "\nYou may want to stash your changes to the file before trying again,"
-  );
-}
+patchFile();
